pyproject.toml
[tool.poetry]
name = "enough_recipes"
version = "0.1.0"
description = ""
authors = ["Robert Elwell <robert.elwell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
boltons = "^21.0.0"
mwparserfromhell = "^0.6.4"
requests = "^2.27.1"
kafka-python = "^2.0.2"
elasticsearch = "^8.2.0"
dj-database-url = "^0.5.0"
mysqlclient = "^2.1.0"
beautifulsoup4 = "^4.11.1"
gunicorn = "^20.1.0"
django-tailwind = "^3.1.1"
django-browser-reload = "^1.6.0"
django-storages = "^1.12.3"
boto3 = "^1.24.7"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
mypy = "^0.950"
types-requests = "^2.27.27"
black = "^22.3.0"
pylint = "^2.14.1"
django-stubs = {extras = ["compatible-mypy"], version = "^1.11.0"}
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "enough_recipes.settings"

[tool.pylint.master]
load-plugins = "pylint_django"
django-settings-module = "enough_recipes.settings"

[[tool.mypy.overrides]]
module = [
    "boltons.*",
    "bs4",
    "dj_database_url",
    "kafka.*"
]
ignore_missing_imports = true
